server-monitor/requirement.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
mailjet-rest==1.3.4
psutil==6.1.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
